# docs/openapi.yaml
openapi: 3.0.3
info:
  title: iag-quality-control
  version: 1.0.0
  description: CUPPA LIMS — samples, tests, certification, CoA, compliance
servers:
  - url: http://localhost:4004
  - url: http://localhost:8080/api/v1/quality-control
paths:
...
  /api/v1/search:
    get:
      summary: Global search across samples, CoA, instruments, certifications
      parameters:
        - in: query
          name: q
          required: true
          schema: { type: string, minLength: 2 }
        - in: query
          name: limit
          schema: { type: integer, default: 20 }
  /api/v1/calendar:
    get:
      summary: Certification, calibration, and audit calendar
      parameters:
        - in: query
          name: from
          schema: { type: string, format: date }
        - in: query
          name: to
          schema: { type: string, format: date }
  /api/v1/analytics/spc:
    get:
      summary: SPC control chart data
      parameters:
        - in: query
          name: metric
          schema: { type: string, enum: [moisture, cup_score], default: moisture }
        - in: query
          name: batch_business_id
          schema: { type: string }
        - in: query
          name: days
          schema: { type: integer, default: 30 }
  /api/v1/reports/day-summary:
    get:
      summary: Day report
      parameters:
        - in: query
          name: date
          schema: { type: string, format: date }
  /api/v1/reports/day-summary/pdf:
    get:
      summary: Day report PDF
      parameters:
        - in: query
          name: date
          schema: { type: string, format: date }
  /api/v1/reports/trends:
    get:
      summary: Daily trend points
      parameters:
        - in: query
          name: days
          schema: { type: integer, default: 7 }
...
  /api/v1/reports/export:
    get:
      summary: CSV export
      parameters:
        - in: query
          name: type
          required: true
          schema: { type: string, enum: [samples] }
...
  /api/v1/samples/{id}/label:
    get:
      summary: Sample barcode label
      parameters:
        - in: query
          name: format
          schema: { type: string, enum: [json, zpl, svg], default: json }
...